requirements.txt
Flask==2.3.3
pywebview==4.4.1
Jinja2==3.1.2
Werkzeug==2.3.7
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3